backend/requirements.txt
python-dotenv
pymongo
requests